daemon/telegram/requirements.txt
apscheduler
httpx
import-expression
mysql-connector-python
pyrogram
python-dotenv
tgcrypto
xmltodict
unicode
python-slugify
